kbuild: Introduce -D__linux_kernel__ to indicate building of the Linux kernel
The __linux__ macro is defined by gcc to indicate that the compiler
is configured for a target running Linux.  However, it also makes
sense to use a bare-metal-targeted compiler to build the kernel,
particularly when cross-building.  Such compilers won't (and
shouldn't) define __linux__.

Code which tests for __linux__ should only appear in code which
will never be built into the kernel, such as build-time tools and
some parts of userspace headers.

In general, it is reasonable for Linux to share some source code
with other projects which could predictably be using __KERNEL__ for
their own purposes.  This patch introduces a separate macro
__linux_kernel__ to help to disambiguate such cases.

Signed-off-by: Dave Martin <dave.martin@linaro.org>