OMAP: Add support to allocate the memory for secure RAM
Allocate the memory to save secure ram context which needs
to be done when MPU is hitting off mode.

The ROM code expects a physical address to this memory
and hence use memblock APIs to reserve this memory as part
of .reserve() callback.

Maximum possible size is allocated and can cater to OMAP3XXX / OMAP4XXX
secure RAM size requirements.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>