plat-omap: mcbsp: Move WAKEUPEN configuration to omap_mcbsp_config
Configure the WAKEUPEN register at the same time we configure the rest of
the McBSP registers.
In case of OMAP2+, if the sysclock has been reconfigured we are going to
disable McBSP for the duration of the clock change, which will reset the
McBSP registers. The WAKEUPEN register need to be configured later, so
the changes will be effective during runtime.

Change-Id: I4085feb2c73a71856aaca47f3ecce06518ffb401
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>