ASoC: OMAP: HDMI: Prevent DSS module from going idle when playing audio
When playing audio through HDMI, the module should be put in
no-idle mode. This is to prevent the DSS_L3_ICLK to be shut down.
Once audio playback is stopped, the module is set back to smart-idle
wakeup-capable. This is as described by OMAP4 TRM.

Also, a omap_hwmod structure is added to hdmi data to have access
to idle-mode configuration.

(AG: updated to new Tomi HDMI patches)

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>
Signed-off-by: Andy Green <andy.green@linaro.org>