syslink: procmgr: initialize attach_count
atomic_cmpmask_and_set() is operating on an uninitialized attach_count
which can pass the cmpmask as if it has been initialized
resulting in the cmpmask_and_set() not initializing the attach_count.
For instance, if the uninitialized attach_count happens to be
0xffffffff, atomic_cmpmask_and_set() will not change the attach_count.
This attach_count is allocated on the heap.

Change-Id: Iaadc3f093287d965ff4d00f8d3f9c63ce596a6d4
Signed-off-by: Bryan Honza <honza@ti.com>