OMAP2+: mailbox: fix lookups for multiple mailboxes
Re-apply the following patch. It was partially reverted by a merge
with mutex addition.
This fix is required because, as expected, the concerned code generates
an error with gcc4.6 (but not with gcc4.5!).

OMAP2+: mailbox: fix lookups for multiple mailboxes

The pointer math in omap_mbox_get() is not quite right, and leads to
passing NULL to strcmp() when searching for an mbox that is not first
in the list.

Convert to using array indexing as is done in all the other functions
which walk the mbox list.

Tested on OMAP2420/n810, OMAP3630/zoom3, OMAP4430/Blaze

Signed-off-by: Kevin Hilman <khilman@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>

Conflicts:

	arch/arm/plat-omap/mailbox.c

Signed-off-by: Sebastien Jan <s-jan@ti.com>