TILER: TCM-SITA: Simplified map handling.
Map used to have copies of the parent area in each slot.  It also
used to have a busy member.  This is now replaced with a pointer
to the parent area which is set to NULL if not occupied.

Also removed reserved member as it was not used.

Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: David Sin <davidsin@ti.com>

TILER: TCM-SiTA: Removed tracking of tiler areas.

Since no method requires knowing the list of tiler areas in
tcm_sita, the tracking of such has been removed.

Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: David Sin <davidsin@ti.com>

TILER: TCM-SiTA: Removed unhandled memory allocation while scanning

Changed scoring to happen incrementally while scanning.  This
avoids creating a list of potential positions that may run out of
memory.

Since we sort positions strictly, we can use an incremental
approach.  Furthermore, if first position selection is enabled,
this reduces the search time.

Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: David Sin <davidsin@ti.com>