hwspinlocks: disable runtime PM
On 3.1 kernel, rpm_idle takes 2.7ms when usage_count=0 (this function
runs under a spinlock with irqs disabled).
The root cause is not known yet, but this is clearly not expected and
is much shorter on 3.0 based kernel.
This causes x2 cpu-load and performances drop on a video decode
use-case using IVAHD (and syslink) on OMAP4. Disabling rpm brings back
load and performance figures in line with k3.0.

While the root-cause is not identified and fixed, keep runtime-pm
disabled for hwspinlocks.

Signed-off-by: Sebastien Jan <s-jan@ti.com>