ipc,msg: always respect MSG_NOERROR
When specifying the MSG_NOERROR flag, receivers can avoid returning error
(E2BIG) and just truncate the message text, if it is too large.

Currently, this logic is only respected when there are already pending
messages in the queue.  Fix this for the case when there are only
receivers waiting for a msg to be sent.  In order for this to work, save
the flags in the msg_receiver struct as it must be used later when doing
the pipeline send.

Also do some pipeline_send() cleanups while at it.

Signed-off-by: Davidlohr Bueso <davidlohr@hp.com>
Cc: Manfred Spraul <manfred@colorfullife.com>
Cc: Aswin Chandramouleeswaran <aswin@hp.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>