x86, vdso: fix vdso2c's special_pages error checking
Stephen Rothwell's compiler did something amazing: it unrolled a loop,
discovered that one iteration of that loop contained an always-true test,
and emitted a warning that will IMO only serve to convince people to
disable the warning.

That bogus warning caused me to wonder what prompted such an absurdity
from his compiler, and I discovered that the code in question was, in
fact, completely wrong -- I was looking things up in the wrong array.

This affects 3.16 as well, but the only effect is to screw up the
error checking a bit.  vdso2c's output is unaffected.

Signed-off-by: Andy Lutomirski <luto@amacapital.net>
Cc: Stephen Rothwell <sfr@canb.auug.org.au>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>